+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.demopizzaapp",
+    package = "com.pizzagustav",
 )
 
 android_resource(
     name = "res",
-    package = "com.demopizzaapp",
+    package = "com.pizzagustav",
     res = "src/main/res",
 )
