+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">demoPizzaApp</string>
+    <string name="app_name">PizzaGustav</string>
 </resources>
