+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'demoPizzaApp'
+rootProject.name = 'PizzaGustav'
 
 include ':app'
