+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "demoPizzaApp",
-  "displayName": "demoPizzaApp"
+  "name": "PizzaGustav",
+  "displayName": "PizzaGustav"
 }
