+++ ios/PizzaGustavTests/PizzaGustavTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface demoPizzaAppTests : XCTestCase
+@interface PizzaGustavTests : XCTestCase
 
 @end
 
-@implementation demoPizzaAppTests
+@implementation PizzaGustavTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
