+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "demoPizzaApp",
+  "name": "PizzaGustav",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
